--- src/test/java/HomeWork16.java
@@ -1,19 +1,0 @@
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import java.time.Duration;
-
-public class HomeWork16 extends BaseTest {
-    @Test
-    public void registrationNavigation()  {
-
-        WebElement registrationLink = getThreadLocal().findElement(By.cssSelector("[id='hel']"));
-        registrationLink.click();
-
-        String registrationUrl = "https://qa.koel.app/registration.php";
-        Assert.assertEquals(getThreadLocal().getCurrentUrl(), registrationUrl);
-
-    }
-}
