+++ src/test/resources/Smoke.xml
@@ -7,7 +7,7 @@
     <test name="Example Test" preserve-order="false">
         <parameter name="baseURL" value="https://qa.koel.app/"/>
         <classes>
-            <class name="HomeWork18">
+            <class name="RandomTests">
                 <methods>
                     <include name="playASongTest" />
                 </methods>
